# RSDWAY/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SFDA DTTS (Rasid) Integration API - Proxy for SOAP web services
servers:
  - url: /api
    description: Base API path
...
paths:
  /external/keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for the logged-in user
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKeyEntry"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a new API key (returned once — store it securely)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "200":
          description: Newly created API key (full key shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedApiKey"
  /external/keys/{id}:
    delete:
      operationId: deleteApiKey
      tags: [api-keys]
      summary: Revoke (delete) an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
  /rasid/accept-dispatch:
    post:
      operationId: acceptDispatch
      tags: [accept]
      summary: Accept all products in a dispatch notification at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptDispatchRequest"
      responses:
        "200":
          description: Accept dispatch response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoapResponse"

  # ── RETURN ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AcceptDispatchRequest:
      type: object
      properties:
        dispatchNotificationId:
          type: string
      required:
        - dispatchNotificationId
...
    SoapResponse:
      type: object
      properties:
        success:
          type: boolean
        rawXml:
          type: ["string", "null"]
        parsed:
          description: Parsed response object
        error:
          type: ["string", "null"]
        notificationId:
          type: ["string", "null"]
      required:
        - success
    ApiKeyEntry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        enabled:
          type: boolean
        keyPreview:
          type: string
          description: Masked key preview (first 12 + last 4 chars)
        createdAt:
          type: string
        lastUsedAt:
          type: ["string", "null"]
      required:
        - id
        - name
        - enabled
        - keyPreview
        - createdAt
    CreateApiKeyRequest:
      type: object
      properties:
        name:
          type: string
          description: A label to identify this key (e.g. "Odoo Production")
      required:
        - name
    CreatedApiKey:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        key:
          type: string
          description: Full API key — shown only once, store it securely
        createdAt:
          type: string
      required:
        - id
        - name
        - key
        - createdAt
...