# RSDWAY/lib/api-zod/src/generated/types/createApiKeyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SFDA DTTS (Rasid) Integration API - Proxy for SOAP web services
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApiKeyRequest {
  /** A label to identify this key (e.g. "Odoo Production") */
  name: string;
}