# RSDWAY/lib/api-zod/src/generated/types/soapResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SFDA DTTS (Rasid) Integration API - Proxy for SOAP web services
 * OpenAPI spec version: 0.1.0
 */

export interface SoapResponse {
  success: boolean;
  /** @nullable */
  rawXml?: string | null;
  /** Parsed response object */
  parsed?: unknown;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  notificationId?: string | null;
}